syncval: Segregate Renderpass support into file
Both renderpass and dynamic rendering
