+++ layers/CMakeLists.txt
@@ -274,6 +274,8 @@ target_sources(vvl PRIVATE
     sync/sync_validation.h
     sync/sync_model.cpp
     sync/sync_model.h
+    sync/sync_renderpass.cpp
+    sync/sync_renderpass.h
     sync/sync_sync_op.cpp
     sync/sync_sync_op.h
     sync/sync_utils.cpp
