+++ layers/sync/sync_sync_op.cpp
@@ -16,6 +16,7 @@
  */
 
 #include "sync/sync_sync_op.h"
+#include "sync/sync_renderpass.h"
 
 // Range generators for to allow event scope filtration to be limited to the top of the resource access traversal pipeline
 //
