+++ layers/sync/sync_validation.h
@@ -29,8 +29,10 @@
 //#include "sync/sync_model.h"
 #include "sync/sync_common.h"
 #include "sync/sync_access_context.h"
+#include "sync/sync_renderpass.h"
 
 class AccessContext;
+struct ClearAttachmentInfo;
 class CommandBufferAccessContext;
 class CommandExecutionContext;
 struct PresentedImage;
@@ -46,44 +48,9 @@ struct SyncEventState;
 class SyncOpBase;
 class SyncOpBeginRenderPass;
 class SyncValidator;
-struct ClearAttachmentInfo;
-
 
 namespace syncval_state {
-
-enum class AttachmentType { kColor, kDepth, kStencil };
-
-struct DynamicRenderingInfo {
-    struct Attachment {
-        const safe_VkRenderingAttachmentInfo &info;
-        std::shared_ptr<const ImageViewState> view;
-        std::shared_ptr<const ImageViewState> resolve_view;
-        ImageRangeGen view_gen;
-        std::optional<ImageRangeGen> resolve_gen;
-        AttachmentType type;
-
-        Attachment(const SyncValidator &state, const safe_VkRenderingAttachmentInfo &info, const AttachmentType type_,
-                   const VkOffset3D &offset, const VkExtent3D &extent);
-
-        SyncStageAccessIndex GetLoadUsage() const;
-        SyncStageAccessIndex GetStoreUsage() const;
-        SyncOrdering GetOrdering() const;
-        Location GetLocation(const Location &loc, uint32_t index = 0) const;
-        bool IsWriteable(const LAST_BOUND_STATE &last_bound_state) const;
-        bool IsValid() const { return view.get(); }
-    };
-
-    // attachments store references to this info, so make sure this doesn't get moved around.
-    DynamicRenderingInfo(const DynamicRenderingInfo &) = delete;
-    DynamicRenderingInfo(DynamicRenderingInfo &&) = delete;
-    DynamicRenderingInfo &operator=(const DynamicRenderingInfo &) = delete;
-    DynamicRenderingInfo &operator=(DynamicRenderingInfo &&) = delete;
-
-    DynamicRenderingInfo(const SyncValidator &state, const VkRenderingInfo &rendering_info);
-    ClearAttachmentInfo GetClearAttachmentInfo(const VkClearAttachment &clear_attachment, const VkClearRect &rect) const;
-    safe_VkRenderingInfo info;
-    std::vector<Attachment> attachments;  // All attachments (with internal typing)
-};
+struct DynamicRenderingInfo;
 
 struct BeginRenderingCmdState {
     BeginRenderingCmdState(std::shared_ptr<const syncval_state::CommandBuffer> &&cb_state_) : cb_state(std::move(cb_state_)) {}
@@ -97,36 +64,6 @@ struct BeginRenderingCmdState {
 VALSTATETRACK_DERIVED_STATE_OBJECT(VkImage, syncval_state::ImageState, IMAGE_STATE)
 VALSTATETRACK_DERIVED_STATE_OBJECT(VkImageView, syncval_state::ImageViewState, IMAGE_VIEW_STATE)
 
-struct ClearAttachmentInfo {
-    using ImageViewState = syncval_state::ImageViewState;
-
-    const ImageViewState *view = nullptr;
-    VkImageAspectFlags aspects_to_clear = {};
-    VkImageSubresourceRange subresource_range{};
-    VkOffset3D offset = {};
-    VkExtent3D extent = {};
-    uint32_t attachment_index = VK_ATTACHMENT_UNUSED;
-    uint32_t subpass = 0;
-
-    static VkImageSubresourceRange RestrictSubresourceRange(const VkClearRect &clear_rect, const ImageViewState &view);
-    static VkImageAspectFlags GetAspectsToClear(VkImageAspectFlags clear_aspect_mask, const ImageViewState &view);
-    ClearAttachmentInfo() = default;
-    ClearAttachmentInfo(const VkClearAttachment &clear_attachment, const VkClearRect &rect, const ImageViewState &view_,
-                        uint32_t attachment_index_ = VK_ATTACHMENT_UNUSED /* renderpass instance only */,
-                        uint32_t subpass_ = 0 /* renderpass instance only */)
-        : view(&view_),
-          aspects_to_clear(GetAspectsToClear(clear_attachment.aspectMask, view_)),
-          subresource_range(RestrictSubresourceRange(rect, view_)),
-          offset(CastTo3D(rect.rect.offset)),
-          extent(CastTo3D(rect.rect.extent)),
-          attachment_index(attachment_index_),
-          subpass(subpass_) {}
-
-    // ClearAttachmentInfo can be invalid for several reasons based on the VkClearAttachment and the rendering
-    // attachment state, including some caught by the constructor.  Consumers *must* check validity before use
-    bool IsValid() const { return view && (aspects_to_clear != 0U) && (subresource_range.layerCount != 0U); }
-    std::string GetSubpassAttachmentText() const;
-};
 
 class AlternateResourceUsage {
   public:
@@ -437,68 +374,6 @@ class SyncEventsContext {
 
   private:
     Map map_;
-};
-
-class RenderPassAccessContext {
-  public:
-    static AttachmentViewGenVector CreateAttachmentViewGen(
-        const VkRect2D &render_area, const std::vector<const syncval_state::ImageViewState *> &attachment_views);
-    RenderPassAccessContext() : rp_state_(nullptr), render_area_(VkRect2D()), current_subpass_(0) {}
-    RenderPassAccessContext(const RENDER_PASS_STATE &rp_state, const VkRect2D &render_area, VkQueueFlags queue_flags,
-                            const std::vector<const syncval_state::ImageViewState *> &attachment_views,
-                            const AccessContext *external_context);
-
-    static bool ValidateLayoutTransitions(const SyncValidationInfo &val_info, const AccessContext &access_context,
-                                          const RENDER_PASS_STATE &rp_state, const VkRect2D &render_area, uint32_t subpass,
-                                          const AttachmentViewGenVector &attachment_views, vvl::Func command);
-
-    static bool ValidateLoadOperation(const SyncValidationInfo &val_info, const AccessContext &access_context,
-                                      const RENDER_PASS_STATE &rp_state, const VkRect2D &render_area, uint32_t subpass,
-                                      const AttachmentViewGenVector &attachment_views, vvl::Func command);
-
-    bool ValidateStoreOperation(const SyncValidationInfo &val_info, vvl::Func command) const;
-    bool ValidateResolveOperations(const SyncValidationInfo &val_info, vvl::Func command) const;
-
-    static void UpdateAttachmentResolveAccess(const RENDER_PASS_STATE &rp_state, const AttachmentViewGenVector &attachment_views,
-                                              uint32_t subpass, const ResourceUsageTag tag, AccessContext access_context);
-
-    static void UpdateAttachmentStoreAccess(const RENDER_PASS_STATE &rp_state, const AttachmentViewGenVector &attachment_views,
-                                            uint32_t subpass, const ResourceUsageTag tag, AccessContext &access_context);
-
-    static void RecordLayoutTransitions(const RENDER_PASS_STATE &rp_state, uint32_t subpass,
-                                        const AttachmentViewGenVector &attachment_views, const ResourceUsageTag tag,
-                                        AccessContext &access_context);
-
-    bool ValidateDrawSubpassAttachment(const CommandExecutionContext &ex_context, const CMD_BUFFER_STATE &cmd_buffer,
-                                       vvl::Func command) const;
-    void RecordDrawSubpassAttachment(const CMD_BUFFER_STATE &cmd_buffer, ResourceUsageTag tag);
-
-    uint32_t GetAttachmentIndex(const VkClearAttachment &clear_attachment) const;
-    ClearAttachmentInfo GetClearAttachmentInfo(const VkClearAttachment &clear_attachment, const VkClearRect &rect) const;
-
-    bool ValidateNextSubpass(const CommandExecutionContext &ex_context, vvl::Func command) const;
-    bool ValidateEndRenderPass(const CommandExecutionContext &ex_context, vvl::Func command) const;
-    bool ValidateFinalSubpassLayoutTransitions(const CommandExecutionContext &ex_context, vvl::Func command) const;
-
-    void RecordLayoutTransitions(ResourceUsageTag tag);
-    void RecordLoadOperations(ResourceUsageTag tag);
-    void RecordBeginRenderPass(ResourceUsageTag tag, ResourceUsageTag load_tag);
-    void RecordNextSubpass(ResourceUsageTag store_tag, ResourceUsageTag barrier_tag, ResourceUsageTag load_tag);
-    void RecordEndRenderPass(AccessContext *external_context, ResourceUsageTag store_tag, ResourceUsageTag barrier_tag);
-
-    AccessContext &CurrentContext() { return subpass_contexts_[current_subpass_]; }
-    const AccessContext &CurrentContext() const { return subpass_contexts_[current_subpass_]; }
-    const std::vector<AccessContext> &GetContexts() const { return subpass_contexts_; }
-    uint32_t GetCurrentSubpass() const { return current_subpass_; }
-    const RENDER_PASS_STATE *GetRenderPassState() const { return rp_state_; }
-    AccessContext *CreateStoreResolveProxy() const;
-
-  private:
-    const RENDER_PASS_STATE *rp_state_;
-    const VkRect2D render_area_;
-    uint32_t current_subpass_;
-    std::vector<AccessContext> subpass_contexts_;
-    AttachmentViewGenVector attachment_views_;
 };
 
 // Command execution context is the base class for command buffer and queue contexts
